requirements.txt
anyio==3.3.1
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==4.1.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cycler==0.11.0
Cython==0.29.23
daphne==3.0.2
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
django-allauth==0.45.0
django-axes==5.27.0
django-ipware==4.0.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.2
gensim==4.1.2
hyperlink==21.0.0
idna==3.2
igraph==0.9.8
incremental==21.3.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.2
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
openai==0.11.1
opencv-contrib-python==4.5.3.56
openpyxl==3.0.9
packaging==21.0
pandas==1.3.2
pandas-stubs==1.2.0.39
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pygame==2.0.2
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.0.0
tk==0.1.0
toml==0.10.2
tornado==6.1
tqdm==4.62.0
traitlets==5.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
zope.interface==5.4.0
